pyproject.toml
[project]
name = "google-maps-mcp-server"
version = "0.1.0"
description = "Production-ready MCP server for Google Maps Platform APIs - Directions, Places, Geocoding, Roads, and Distance Matrix"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
    { name = "Etty Sekhon", email = "etty.sekhon@gmail.com" }
]
keywords = [
    "mcp",
    "model-context-protocol",
    "google-maps",
    "ai-agents",
    "llm-tools",
    "adk"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mcp>=1.22.0",
    "googlemaps>=4.10.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "structlog>=25.5.0",
    "httpx>=0.28.1",
    "google-adk>=1.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-timeout>=2.4.0",
    "pytest-mock>=3.15.0",
    "black>=25.11.0",
    "ruff>=0.14.6",
    "mypy>=1.18.2",
    "pre-commit>=4.5.0",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=0.19.0",
]

[project.urls]
Homepage = "https://github.com/ettysekhon/google-maps-mcp-server"
Documentation = "https://github.com/ettysekhon/google-maps-mcp-server#readme"
Repository = "https://github.com/ettysekhon/google-maps-mcp-server"
Issues = "https://github.com/ettysekhon/google-maps-mcp-server/issues"
Changelog = "https://github.com/ettysekhon/google-maps-mcp-server/blob/main/CHANGELOG.md"

[project.scripts]
google-maps-mcp-server = "google_maps_mcp_server.server:main"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "black>=25.11.0",
    "ruff>=0.14.6",
    "mypy>=1.18.2",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312", "py313", "py314"]

[tool.mypy]
python_version = "3.14"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["googlemaps", "googlemaps.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["mcp.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["tenacity"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["google_maps_mcp_server.*", "tests.*", "examples.*"]
disable_error_code = ["misc", "no-untyped-call"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "--cov=src/google_maps_mcp_server",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "*/conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
